requirements.txt
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=0.1.0
anthropic>=0.40.0
python-dotenv>=1.0.0
yfinance>=0.2.30
rich>=13.0.0
python-pptx>=0.6.21
requests>=2.31.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
